PlayerFactory initial
