+++ Source/ACE.Server/Command/Handlers/AdminCommands.cs
@@ -1399,7 +1399,7 @@ namespace ACE.Server.Command.Handlers
 
             weenie.Type = (int)session.Player.WeenieType;
 
-            var player = new Player(weenie, guid, session);
+            var player = new Player(weenie, guid, session.Id);
             player.Location = session.Player.Location;
 
             player.Character.CharacterOptions1 = session.Player.Character.CharacterOptions1;
