+++ Source/ACE.Server/WorldObjects/Admin.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using ACE.Database.Models.Shard;
 using ACE.Database.Models.World;
 using ACE.Entity;
-using ACE.Entity.Enum;
 using ACE.Server.Network;
 
 namespace ACE.Server.WorldObjects
@@ -13,7 +12,7 @@ namespace ACE.Server.WorldObjects
         /// <summary>
         /// A new biota be created taking all of its values from weenie.
         /// </summary>
-        public Admin(Weenie weenie, ObjectGuid guid, Session session) : base(weenie, guid, session)
+        public Admin(Weenie weenie, ObjectGuid guid, uint accountId) : base(weenie, guid, accountId)
         {
             SetEphemeralValues();
         }
