+++ Source/ACE.Server/WorldObjects/Sentinel.cs
@@ -4,7 +4,6 @@ using ACE.Database.Models.Shard;
 using ACE.Database.Models.World;
 using ACE.Entity;
 using ACE.Entity.Enum;
-using ACE.Entity.Enum.Properties;
 using ACE.Server.Network;
 
 namespace ACE.Server.WorldObjects
@@ -14,7 +13,7 @@ namespace ACE.Server.WorldObjects
         /// <summary>
         /// A new biota be created taking all of its values from weenie.
         /// </summary>
-        public Sentinel(Weenie weenie, ObjectGuid guid, Session session) : base(weenie, guid, session)
+        public Sentinel(Weenie weenie, ObjectGuid guid, uint accountId) : base(weenie, guid, accountId)
         {
             Character.IsPlussed = true;
